requirements.txt
attrs==25.4.0
python-dotenv==1.2.1
web3==7.14.0
xrpl-py==4.3.1
